PULL_REQUESTED: s5p-mfc: Support multiple firmware sub-versions
For MFC firmwares, improved versions with bug fixes and
feature additions are released keeping the firmware version
including major and minor number same. The issue came with
the release of a new MFCv6 firmware with an interface change.
This patch adds the support of accepting multiple firmware
binaries for every version with the driver trying to load
firmwares starting from latest. This ensures full backward
compatibility regardless of which firmware version and kernel
version is used.

Change-Id: I67bd62143a7b61b737ad22862ffd81b6bfb68c9e
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>