HACK: vb2: Dont return error if buffer is in use
While freeing buffers, VB2 now returns error if the buffer
is in use. When application exports the buffer and gets
an fd, the refcount is incremented and so the buffer cant be
freed unless the application releases the fd. Currently the
chrome app is not doing it causing resolution change scenarios
to fail. Chrome player app should close the fd exported once
the resolution change event is raised by driver. This check is
recently added in VB2.
Keeping this hack till the fix is done in chrome app.

Signed-off-by: Arun Kumar K <arun.kk@samsung.com>